Get the tile sources for Copernicus GLO-30, ALOS World 3D-30m, NASADEM HGT v001 DTM using the planetary computer STAC catalogue.

Get the tile sources for Copernicus GLO-30, ALOS World 3D-30m, NASADEM HGT v001 DTM using the planetary computer STAC catalogue.

Usage

mpc_dtm_src(
  aoi,
  collection = c("cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem"),
  vsicurl = TRUE
)

Arguments

aoi

a spatial extent which can be provided as a numeric bounding box or using a spatial object from either the terra, sf or stars packages.

collection

Any of: "cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem".

vsicurl

logical with default TRUE. If TRUE the '/vsicurl/' prefix is appended to returned urls which increases read speed for gdal. see https://gdal.org/user/virtual_file_systems.html#vsicurl-http-https-ftp-files-random-access for more information.

Value

A character vector of url paths.

Examples

# define bbox with numeric
mpc_dtm_src(c( -121.772, 45.321, -121.611, 45.419))
#> [1] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP30_hh/Copernicus_DSM_COG_10_N45_00_W122_00_DEM.tif"

if (requireNamespace("sf", quietly = TRUE)){
nc_sf <- system.file("gpkg", "nc.gpkg", package = "sf") |>
sf::read_sf()
mpc_dtm_src(nc_sf, collection = "alos-dem")
}
#>  [1] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W085_DSM.tif"
#>  [2] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W084_DSM.tif"
#>  [3] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W083_DSM.tif"
#>  [4] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W082_DSM.tif"
#>  [5] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W081_DSM.tif"
#>  [6] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W080_DSM.tif"
#>  [7] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W079_DSM.tif"
#>  [8] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W078_DSM.tif"
#>  [9] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W077_DSM.tif"
#> [10] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W076_DSM.tif"
#> [11] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W085_DSM.tif"
#> [12] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W084_DSM.tif"
#> [13] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W083_DSM.tif"
#> [14] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W082_DSM.tif"
#> [15] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W081_DSM.tif"
#> [16] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W080_DSM.tif"
#> [17] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W079_DSM.tif"
#> [18] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W078_DSM.tif"
#> [19] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W077_DSM.tif"
#> [20] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W076_DSM.tif"
#> [21] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W085_DSM.tif"
#> [22] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W084_DSM.tif"
#> [23] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W083_DSM.tif"
#> [24] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W082_DSM.tif"
#> [25] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W081_DSM.tif"
#> [26] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W080_DSM.tif"
#> [27] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W079_DSM.tif"
#> [28] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W078_DSM.tif"
#> [29] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W077_DSM.tif"
#> [30] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W085_DSM.tif"
#> [31] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W084_DSM.tif"
#> [32] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W083_DSM.tif"
#> [33] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W082_DSM.tif"
#> [34] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W081_DSM.tif"
#> [35] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W080_DSM.tif"
#> [36] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W079_DSM.tif"
#> [37] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W078_DSM.tif"

if (requireNamespace("terra", quietly = TRUE)){
fr <- system.file("ex/elev.tif", package="terra") |>
terra::rast()
mpc_dtm_src(nc_sf, collection = "cop-dem-glo-90")
}
#>  [1] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W085_00_DEM.tif"
#>  [2] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W084_00_DEM.tif"
#>  [3] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W083_00_DEM.tif"
#>  [4] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W082_00_DEM.tif"
#>  [5] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W081_00_DEM.tif"
#>  [6] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W080_00_DEM.tif"
#>  [7] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W079_00_DEM.tif"
#>  [8] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W078_00_DEM.tif"
#>  [9] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W077_00_DEM.tif"
#> [10] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W076_00_DEM.tif"
#> [11] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W085_00_DEM.tif"
#> [12] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W084_00_DEM.tif"
#> [13] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W083_00_DEM.tif"
#> [14] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W082_00_DEM.tif"
#> [15] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W081_00_DEM.tif"
#> [16] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W080_00_DEM.tif"
#> [17] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W079_00_DEM.tif"
#> [18] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W078_00_DEM.tif"
#> [19] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W077_00_DEM.tif"
#> [20] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W076_00_DEM.tif"
#> [21] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W085_00_DEM.tif"
#> [22] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W084_00_DEM.tif"
#> [23] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W083_00_DEM.tif"
#> [24] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W082_00_DEM.tif"
#> [25] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W081_00_DEM.tif"
#> [26] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W080_00_DEM.tif"
#> [27] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W079_00_DEM.tif"
#> [28] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W078_00_DEM.tif"
#> [29] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W077_00_DEM.tif"
#> [30] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W085_00_DEM.tif"
#> [31] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W084_00_DEM.tif"
#> [32] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W083_00_DEM.tif"
#> [33] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W082_00_DEM.tif"
#> [34] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W081_00_DEM.tif"
#> [35] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W080_00_DEM.tif"
#> [36] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W079_00_DEM.tif"
#> [37] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W078_00_DEM.tif"