Create an rstac compatible search boundary box

Create an rstac compatible search boundary box

Usage

rstac_bbox(x)

# S3 method for class 'sf'
rstac_bbox(x)

# S3 method for class 'sfc'
rstac_bbox(x)

# S3 method for class 'stars'
rstac_bbox(x)

# S3 method for class 'SpatRaster'
rstac_bbox(x)

# S3 method for class 'SpatVector'
rstac_bbox(x)

# S3 method for class 'numeric'
rstac_bbox(x)

Arguments

x

A numeric (length 4) to define the spatial bounds of the STAC API query or one of the following rspatial objects: sf, sfc, stars, SpatVector, SpatRaster

Value

A numeric of length 4 defining the xmin, ymin, xmax ymax coordinates of the search extent in WGS84 (EPSG:4326).

Examples

mpc_dtm_src(c(-121.772, 45.321, -121.611, 45.419))
#> [1] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP30_hh/Copernicus_DSM_COG_10_N45_00_W122_00_DEM.tif?st=2024-10-10T16%3A16%3A37Z&se=2024-10-11T17%3A01%3A37Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T09%3A18%3A22Z&ske=2024-10-18T09%3A18%3A22Z&sks=b&skv=2024-05-04&sig=rp89so8JluXfgaEMY%2FMGVNOC3oLonAF9nqmCUQiF0vg%3D"