Get the Digital Elevation Model (DEM) tile url sources

Get the Digital Elevation Model (DEM) tile url sources from the Microsoft Planetary Computer STAC catalogue.

Usage

mpc_dtm_src(
  aoi,
  collection = c("cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem"),
  vsicurl = TRUE,
  sign_items = TRUE,
  subscription_key = Sys.getenv("rsi_pc_key")
)

Arguments

aoi

a spatial extent which can be provided as a numeric bounding box or using a spatial object from either the terra, sf or stars packages.

collection

A character. one of: "cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem".

vsicurl

logical with default TRUE. If TRUE the '/vsicurl/' prefix is appended to returned urls which increases read speed for gdal. see https://gdal.org/user/virtual_file_systems.html#vsicurl-http-https-ftp-files-random-access for more information.

sign_items

logical with default TRUE. If TRUE the STAC items are signed this is now required for the planetary computer API.

subscription_key

A character. The subscription key for the planetary computer API. This can be set as an environment variable 'rsi_pc_key' ( the default for the rsi package) or 'MPC_TOKEN' (the default for the sits package). It isn't essential to set this token but it should improve the performance of the API. See sign_planetary_computer

Value

A character vector of url paths.

Examples

# define bbox with numeric
mpc_dtm_src(c(-121.772, 45.321, -121.611, 45.419))
#> [1] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP30_hh/Copernicus_DSM_COG_10_N45_00_W122_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"

if (requireNamespace("sf", quietly = TRUE)) {
  nc_sf <- system.file("gpkg", "nc.gpkg", package = "sf") |>
    sf::read_sf()
  mpc_dtm_src(nc_sf, collection = "alos-dem")
}
#>  [1] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W085_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [2] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W084_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [3] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W083_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [4] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W082_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [5] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W081_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [6] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W080_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [7] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W079_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [8] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W078_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#>  [9] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W077_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [10] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N036W076_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [11] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W085_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [12] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W084_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [13] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W083_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [14] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W082_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [15] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W081_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [16] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W080_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [17] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W079_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [18] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W078_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [19] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W077_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [20] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N035W076_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [21] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W085_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [22] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W084_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [23] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W083_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [24] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W082_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [25] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W081_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [26] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W080_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [27] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W079_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [28] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W078_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [29] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N034W077_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [30] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W085_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [31] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W084_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [32] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W083_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [33] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W082_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [34] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W081_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [35] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W080_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [36] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W079_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"
#> [37] "/vsicurl/https://ai4edataeuwest.blob.core.windows.net/alos-dem/AW3D30_global/ALPSMLC30_N033W078_DSM.tif?st=2024-10-10T22%3A09%3A32Z&se=2024-10-11T22%3A54%3A32Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T18%3A05%3A07Z&ske=2024-10-18T18%3A05%3A07Z&sks=b&skv=2024-05-04&sig=KYbDsKBiFj1ArIwkM8l4x%2FIXPwkr7YV6kYGC9Xc%2FYFs%3D"

if (requireNamespace("terra", quietly = TRUE)) {
  fr <- system.file("ex/elev.tif", package = "terra") |>
    terra::rast()
  mpc_dtm_src(nc_sf, collection = "cop-dem-glo-90")
}
#>  [1] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W085_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [2] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W084_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [3] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W083_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [4] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W082_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [5] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W081_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [6] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W080_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [7] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W079_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [8] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W078_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#>  [9] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W077_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [10] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N36_00_W076_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [11] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W085_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [12] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W084_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [13] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W083_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [14] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W082_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [15] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W081_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [16] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W080_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [17] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W079_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [18] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W078_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [19] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W077_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [20] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N35_00_W076_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [21] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W085_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [22] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W084_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [23] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W083_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [24] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W082_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [25] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W081_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [26] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W080_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [27] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W079_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [28] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W078_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [29] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N34_00_W077_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [30] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W085_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [31] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W084_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [32] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W083_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [33] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W082_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [34] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W081_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [35] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W080_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [36] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W079_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"
#> [37] "/vsicurl/https://elevationeuwest.blob.core.windows.net/copernicus-dem/COP90_hh/Copernicus_DSM_COG_30_N33_00_W078_00_DEM.tif?st=2024-10-10T22%3A09%3A26Z&se=2024-10-11T22%3A54%3A26Z&sp=rl&sv=2024-05-04&sr=c&skoid=9c8ff44a-6a2c-4dfb-b298-1c9212f64d9a&sktid=72f988bf-86f1-41af-91ab-2d7cd011db47&skt=2024-10-11T21%3A27%3A35Z&ske=2024-10-18T21%3A27%3A35Z&sks=b&skv=2024-05-04&sig=ldqP6ztf6dGXYZT7%2Fd2Q%2BKp5nFBuVnSxzVkcrILYDyg%3D"